\name{oneStepSim}
\alias{oneStepSim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
One-step function for data simulation
}
\description{
This function simulates pure and mixed cell-type expression data for two cell types, namely T and N. The simulation is based on multivariate normal (MVN) distribution.
}
\usage{
oneStepSim(n.samp, mu.T, mu.N, Sigma.T=NULL, Sigma.N=NULL, prop.T="seq", 
           # structure for Sigma.T
           block.size, rho, dd=NULL, str.type="interchangeable",
           # selected genes to add structure
           select.gene="first")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n.samp}{number of samples to simulate for each cell type.}
  \item{mu.T}{mean vector for cell type T.}
  \item{mu.N}{mean vector for cell type N.}
  \item{Sigma.T}{covariance matrix for cell type T. If \code{NULL}, covariance matrix with special structure will be generated by \code{generate.Sigma()}. See also \code{generate.Sigma()}.}
  \item{Sigma.N}{covariance matrix for cell type N. If \code{NULL}, identity matrix will be generated.}
  \item{prop.T}{a vector of mixing proportions for cell type T. Options: \code{"seq"} (default), \code{"unif"}, or a customized vector of length \code{n.samp}. If \code{"seq"}, \code{prop.T = seq(0, 1, length=n.samp)}; if \code{"unif"}, \code{prop.T = runif(n.samp, 0, 1)}. And proportions for cell type N are \code{1-prop.T}.} 
  \item{block.size}{see \code{generate.Sigma()}.}
  \item{rho}{see \code{generate.Sigma()}.}
  \item{dd}{see \code{generate.Sigma()}.}
  \item{str.type}{see \code{generate.Sigma()}.}
  \item{select.gene}{select which genes to add special structures on. Note that \code{mu.T} and \code{mu.N} should have the same order of genes. Options: \code{"first"} (default) means to select from the first gene in the mean vector; \code{"random"} means to randomly select genes; \code{"DEG"} means to select differentially expressed genes (according to \code{order(abs(mu.T-mu.N),decreasing=TRUE)}).}
}
\details{
The function takes the mean expression of each cell type at the log2-transformed scale and mixes the cell types in the raw scale (i.e. 2^x). If covariance matrices are not given, by default, the function generates an identity matrix for \code{Sigma.N}; and \code{Sigma.T} can be customized using corresponding arguments. Mixing proportion of cell type T is required; cell type N takes the complementary proportion. Pure cell-type samples are simulated from MVN; after mixing, the mixed samples are log2-transformed.

All MVN simulated expression data are forced to be non-negative by replacing negative values with zero. All simulated samples are quantile normalized by \code{normalize.quantiles()} from the \code{preprocessCore} package.
}
\value{
A list with the following components:
  \item{m.gene}{number of genes.}
  \item{Sigma.T}{covariance matrix of cell type T.}
  \item{Sigma.N}{covariance matrix of cell type N.}
  \item{true.str.T}{true structure in generated \code{Sigma.T}. A matrix with 0 (no structure) and 1 (an edge). If \code{Sigma.T} is given, it returns \code{NULL}.}
  \item{expr.pure.T}{expression matrix for pure cell type T.}
  \item{expr.pure.N}{expression matrix for pure cell type N.}
  \item{expr.mixed}{expression matrix for mixed cell types.}
  \item{prop.T}{true proportions of cell type T in mixed samples.}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See \code{\link[simDeNet]{sim.expr}} and \code{\link[simDeNet]{generate.Sigma}} for more intermediate details.

\code{\link[mvtnorm]{rmvnorm}} from the \code{mvtnorm} package is used for simulate random MVN samples.

\code{\link[preprocessCore]{normalize.quantiles}} from the \code{preprocessCore} package is used for quantile normalization.
}
\examples{
## load cell type data
set.seed(999)
data("celltype")
mu.T <- expr[,ctab$Fastq_file_name[which(ctab$X3_letter_code=="ASM")]]
mu.N <- expr[,ctab$Fastq_file_name[which(ctab$X3_letter_code=="AEC")]]

## number of samples to simulate
n.samp <- 5

## parameters for correlation design of cell type T
rho <- c(0.9,0.8,0.7)
block.size <- c(5,10,15)
str.type <- c("interchangeable","decaying","star")
select.gene <- "DEG"

## mixing proportion of cell type T
prop.T <- runif(n.samp)

## one-step simulation
out.oneStepSim <- oneStepSim(n.samp, mu.T, mu.N, Sigma.T=NULL, Sigma.N=NULL, prop.T=prop.T, 
                             # structure for Sigma.T
                             dd=NULL, rho=rho, block.size=block.size, str.type=str.type,
                             # selected genes to add structure
                             select.gene=select.gene)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ datagen }% use one of  RShowDoc("KEYWORDS")
